tests/wpt/tests/html/canvas/element/line-styles/2d.line.fill.noop.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.line.fill.noop</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.line.fill.noop</h1>
<p class="desc">Filling a line draws nothing</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("Filling a line draws nothing");
_addTest(function(canvas, ctx) {

  ctx.fillStyle = '#0f0';
  ctx.fillRect(0, 0, 100, 50);
  ctx.fillStyle = '#f00';
  ctx.lineWidth = 20;
  ctx.beginPath();
  ctx.moveTo(10, 20);
  ctx.lineTo(90, 30);
  ctx.fill();
  _assertPixel(canvas, 50,24, 0,255,0,255);
  _assertPixel(canvas, 50,25, 0,255,0,255);
  _assertPixel(canvas, 50,26, 0,255,0,255);

});
</script>
